+++ src/push_socket/push_frame_serialize.cpp
@@ -1,5 +1,4 @@
 #include "tigerapi/push_socket/push_frame_serialize.h"
-#include <boost/endian/conversion.hpp>
 
 std::vector<unsigned char> TIGER_API::PushFrameEncoder::encode_frame(const std::string& packed_frame)
 {
@@ -16,20 +15,20 @@ std::vector<unsigned char> TIGER_API::PushFrameEncoder::encode_frame(const std::
 	while (pack_size) 
 	{
 		unsigned char header_byte = 0x80 | bits;
-		pack_array.push_back(/*boost::endian::native_to_big(header_byte)*/header_byte);
+		pack_array.push_back(header_byte);
 		bits = pack_size & 0x7F;
 		pack_size >>= 7;
 	}
 
-	pack_array.push_back(/*boost::endian::native_to_big(bits)*/bits);
+	pack_array.push_back(bits);
 	pack_array.insert(pack_array.end(), packed_frame.begin(), packed_frame.end());
 	return pack_array;
 }
 
 bool TIGER_API::PushFrameDecoder::push_byte(unsigned char value)
 {
-	//逐个字节压入缓冲区(使用大端序列)
-	frame_header_buffer_.push_back(/*boost::endian::native_to_big(value)*/value);
+	//逐个字节压入缓冲区
+	frame_header_buffer_.push_back(value);
 	if (!(value & 0x80))
 	{
 		//完整读取到包长度
@@ -42,12 +41,12 @@ bool TIGER_API::PushFrameDecoder::push_byte(unsigned char value)
 	return false;
 }
 
-uint64_t TIGER_API::PushFrameDecoder::get_frame_size() const
+uint32_t TIGER_API::PushFrameDecoder::get_frame_size() const
 {
 	return frame_len_;
 }
 
-uint64_t TIGER_API::PushFrameDecoder::decode_varint()
+uint32_t TIGER_API::PushFrameDecoder::decode_varint()
 {
 	const uint32_t mask = (1U << 32) - 1;  // 32位掩码
 	uint32_t result = 0;
